+++ src/main/java/carTravel/repository/issue/IssueRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.sql.Time;
 
-public  interface  IssueRepository extends JpaRepository<Issue, Long> {
+public  interface  IssueRepository extends JpaRepository<Issue, Integer> {
     Issue findIssuesByIsDoneIsFalseOrderByPlanDateEndDesc();
     Issue findIssuesByIsDoneIsTrueOrderByPlanDateEndDesc();
-    Issue findIssuesByExecutorId(Long executorId);
+    Issue findIssuesByExecutorId(Integer executorId);
     void deleteAllByPlanDateEndBefore(Time dateTime);
 
   }
