+++ src/main/java/carTravel/repository/task/TaskRepository.java
@@ -3,5 +3,5 @@ package carTravel.repository.task;
 import carTravel.entity.Tasks;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepository  extends JpaRepository<Tasks, Long> {
+public interface TaskRepository  extends JpaRepository<Tasks, Integer> {
 }
